backend/requirements.txt
django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.6
pymysql>=1.1
httpx>=0.27
python-dotenv>=1.0
PyYAML>=6.0
jsonpath-ng>=1.6
paramiko>=3.4
